Fix headers for klee-uclibc porse threading support

+++ include/klee/runtime/kpr/list-types.h
@@ -1,12 +1,10 @@
 #ifndef KPR_LIST_TYPES
 #define KPR_LIST_TYPES
 
-#include <stddef.h>
-
 typedef struct {
   void* tail;
   void* head;
-  size_t size;
+  unsigned long size;
 } kpr_list;
 
 #define KPR_LIST_INITIALIZER { NULL, NULL, 0 }
